+++ src/Tests/Tests.h
@@ -4,5 +4,5 @@ namespace Tests {
 	void generador();
 	void hash();
 	void enc();
-	void enc_x(const unsigned long long x);
+	void enc_x(const size_t x);
 }
